SY-4532: Move tracked rack/device cleanup into a HardwareCase mixin
TestCase is generic conductor infrastructure; rack/device tracking is a
hardware-domain concern. It now lives in framework/hardware_case.py as a
lifecycle mixin in the SimulatorCase/SimDaqCase mold, and TestCase.teardown
reverts to a plain hook. SimulatorCase, NIRackCase, and NoDevice mix it in.

create_test_ni_rack also derives the device key from the device name (uuid5)
instead of taking caller-supplied magic UUIDs: nothing depends on the specific
values, and a name-derived key keeps reruns after a killed teardown upserting
the same device instead of duplicating it. NoDevice's inline device key gets
the same treatment.
