+++ integration/tests/console/task/no_device.py
@@ -9,13 +9,15 @@
 
 import platform
 import random
+import uuid
 
 import synnax as sy
 from console.case import ConsoleCase
 from console.task.analog_read import AnalogRead
+from framework.hardware_case import HardwareCase
 
 
-class NoDevice(ConsoleCase):
+class NoDevice(HardwareCase, ConsoleCase):
     """
     Verify status message/level when attempting to
     configure and run a task with selected devices
@@ -53,7 +55,7 @@ class NoDevice(ConsoleCase):
         self.create_test_devices(
             [
                 sy.ni.Device(
-                    key="a0e37b26-5401-413e-8e65-c7ad9d9afd70",
+                    key=str(uuid.uuid5(uuid.NAMESPACE_DNS, dev_name)),
                     rack=rack.key,
                     name=dev_name,
                     model=dev_name,
